#include gimpwidgets.h for gimp_spin_button_new().
2005-04-18  Manish Singh  <yosh@gimp.org>

        * libgimpwidgets/gimpsizeentry.c: #include gimpwidgets.h for
        gimp_spin_button_new().

        * plug-ins/jpeg/jpeg-load.c
        * plug-ins/jpeg/jpeg-save.c
        * plug-ins/metadata/interface.c
        * plug-ins/metadata/xmp-parse.c: 64-bit safety for some of the
        debugging g_prints.

+++ plug-ins/jpeg/jpeg-load.c
@@ -296,7 +296,7 @@ load_image (const gchar *filename,
         {
           /* FIXME: handle EXIF here once we don't use libexif anymore */
           g_print ("jpeg-load: found EXIF block (%d bytes)\n",
-                   marker->data_length - sizeof (JPEG_APP_HEADER_EXIF));
+                   (int) (marker->data_length - sizeof (JPEG_APP_HEADER_EXIF)));
           /* Note: maybe split the loop to ensure that the EXIF block is */
           /*       always parsed before any XMP packet */
         }
@@ -309,7 +309,7 @@ load_image (const gchar *filename,
           gchar     *xmp_packet;
 
           g_print ("jpeg-load: found XMP packet (%d bytes)\n",
-                   marker->data_length - sizeof (JPEG_APP_HEADER_XMP));
+                   (int) (marker->data_length - sizeof (JPEG_APP_HEADER_XMP)));
           xmp_packet = g_strndup (marker->data + sizeof (JPEG_APP_HEADER_XMP),
                                   marker->data_length - sizeof (JPEG_APP_HEADER_XMP));
           /* FIXME: running this through the PDB is not very efficient */
